+++ bmcd/src/api.rs
@@ -16,6 +16,7 @@ pub mod legacy;
 pub mod metrics;
 pub mod openapi;
 pub mod paths;
+pub mod responses;
 use self::into_legacy_response::{LegacyResponse, LegacyResult};
 use crate::hal::NodeId;
 use actix_web::web;